requirements.txt
coverage==7.6.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1